# Front End with code/Front End - React/ai-sdr-frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI SDR Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/follow-up/plans": {
      "post": {
        "tags": [
          "follow-up"
        ],
        "summary": "Create Plan",
        "operationId": "create_plan_follow_up_plans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpPlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "follow-up"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_follow_up_plans_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/follow-up/plans/{plan_id}": {
      "get": {
        "tags": [
          "follow-up"
        ],
        "summary": "Get Plan",
        "operationId": "get_plan_follow_up_plans__plan_id__get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/follow-up/plans/{plan_id}/approve": {
      "post": {
        "tags": [
          "follow-up"
        ],
        "summary": "Approve Plan",
        "operationId": "approve_plan_follow_up_plans__plan_id__approve_post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/follow-up/scheduler/run-due": {
      "post": {
        "tags": [
          "follow-up"
        ],
        "summary": "Process Due",
        "operationId": "process_due_follow_up_scheduler_run_due_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FollowUpPlan"
                  },
                  "title": "Response Process Due Follow Up Scheduler Run Due Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/follow-up/scheduler/run-plan/{plan_id}": {
      "post": {
        "tags": [
          "follow-up"
        ],
        "summary": "Process Plan",
        "operationId": "process_plan_follow_up_scheduler_run_plan__plan_id__post",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpPlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FollowUpApprovalRequest": {
        "properties": {
          "approved_by": {
            "type": "string",
            "title": "Approved By"
          }
        },
        "type": "object",
        "required": [
          "approved_by"
        ],
        "title": "FollowUpApprovalRequest"
      },
      "FollowUpListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FollowUpPlan"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "FollowUpListResponse"
      },
      "FollowUpPlan": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id"
          },
          "contact_id": {
            "type": "string",
            "title": "Contact Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_approval",
              "active",
              "paused",
              "stopped",
              "completed"
            ],
            "title": "Status",
            "default": "pending_approval"
          },
          "touches": {
            "items": {
              "$ref": "#/components/schemas/FollowUpTouch"
            },
            "type": "array",
            "title": "Touches"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          },
          "scheduler_backend": {
            "type": "string",
            "enum": [
              "local",
              "temporal"
            ],
            "title": "Scheduler Backend",
            "default": "local"
          },
          "temporal_workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporal Workflow Id"
          },
          "temporal_schedule_status": {
            "type": "string",
            "enum": [
              "not_started",
              "started",
              "failed"
            ],
            "title": "Temporal Schedule Status",
            "default": "not_started"
          },
          "temporal_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temporal Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "campaign_id",
          "contact_id"
        ],
        "title": "FollowUpPlan"
      },
      "FollowUpPlanRequest": {
        "properties": {
          "campaign_id": {
            "type": "string",
            "title": "Campaign Id"
          },
          "delays_hours": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Delays Hours"
          },
          "require_approval": {
            "type": "boolean",
            "title": "Require Approval",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "campaign_id"
        ],
        "title": "FollowUpPlanRequest"
      },
...
      "FollowUpTouch": {
        "properties": {
          "touch_id": {
            "type": "string",
            "title": "Touch Id"
          },
          "sequence_order": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sequence Order"
          },
          "delay_hours": {
            "type": "integer",
            "maximum": 2160.0,
            "minimum": 1.0,
            "title": "Delay Hours"
          },
          "channel": {
            "type": "string",
            "enum": [
              "email",
              "linkedin",
              "phone",
              "sms",
              "whatsapp"
            ],
            "title": "Channel",
            "default": "email"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "title": "Scheduled At"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending_approval",
              "approved",
              "queued",
              "sent",
              "human_task",
              "failed",
              "suppressed"
            ],
            "title": "Status",
            "default": "pending_approval"
          },
          "provider_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Message Id"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          }
        },
        "type": "object",
        "required": [
          "sequence_order",
          "delay_hours",
          "body",
          "scheduled_at"
        ],
        "title": "FollowUpTouch"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}